Moved drepr extension under export extension module, renamed drepr script as main.py
